server/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
chromadb
langchain
langchain-community
tiktoken
openai
google-generativeai
python-dotenv
requests
pandas
